+++ src/aui-layout/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2033](https://issues.liferay.com/browse/AUI-2033) Change add/resize column icon
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)